+++ src/odf/dataset/src/services/metadata_chain_visitors.rs
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+use std::collections::{HashSet, VecDeque};
 use std::error::Error;
 use std::marker::PhantomData;
 
@@ -15,6 +16,7 @@ use odf_metadata::{
     AddData,
     AddPushSource,
     AsTypedBlock,
+    DatasetKind,
     ExecuteTransform,
     IntoDataStreamBlock,
     MetadataBlock,
@@ -69,7 +71,7 @@ macro_rules! typed_search_single_typed_block_visitor_impl {
         pub struct $name {}
 
         impl $name {
-            #[allow(clippy::new_ret_no_self)]
+            #[expect(clippy::new_ret_no_self)]
             pub fn new() -> SearchSingleTypedBlockVisitor<$event_struct> {
                 SearchSingleTypedBlockVisitor::new($block_type_flags)
             }
@@ -79,14 +81,29 @@ macro_rules! typed_search_single_typed_block_visitor_impl {
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+macro_rules! typed_kind_based_search_single_typed_block_visitor_impl {
+    ($expected_kind:expr, $name:ident, $event_struct:ty, $block_type_flags:expr) => {
+        pub struct $name {}
+
+        impl $name {
+            #[expect(clippy::new_ret_no_self)]
+            pub fn new(
+                actual_dataset_kind: DatasetKind,
+            ) -> DatasetKindBasedVisitor<SearchSingleTypedBlockVisitor<$event_struct>, Infallible>
+            {
+                let visitor = SearchSingleTypedBlockVisitor::new($block_type_flags);
+                DatasetKindBasedVisitor::new($expected_kind, actual_dataset_kind, visitor)
+            }
+        }
+    };
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 typed_search_single_typed_block_visitor_impl!(SearchSetVocabVisitor, SetVocab, Flag::SET_VOCAB);
 typed_search_single_typed_block_visitor_impl!(SearchSeedVisitor, Seed, Flag::SEED);
-typed_search_single_typed_block_visitor_impl!(
-    SearchSetPollingSourceVisitor,
-    SetPollingSource,
-    Flag::SET_POLLING_SOURCE
-);
-typed_search_single_typed_block_visitor_impl!(
+typed_kind_based_search_single_typed_block_visitor_impl!(
+    DatasetKind::Derivative,
     SearchSetTransformVisitor,
     SetTransform,
     Flag::SET_TRANSFORM
@@ -96,7 +113,8 @@ typed_search_single_typed_block_visitor_impl!(
     SetDataSchema,
     Flag::SET_DATA_SCHEMA
 );
-typed_search_single_typed_block_visitor_impl!(
+typed_kind_based_search_single_typed_block_visitor_impl!(
+    DatasetKind::Derivative,
     SearchExecuteTransformVisitor,
     ExecuteTransform,
     Flag::EXECUTE_TRANSFORM
@@ -112,7 +130,12 @@ typed_search_single_typed_block_visitor_impl!(
     SetLicense,
     Flag::SET_LICENSE
 );
-typed_search_single_typed_block_visitor_impl!(SearchAddDataVisitor, AddData, Flag::ADD_DATA);
+typed_kind_based_search_single_typed_block_visitor_impl!(
+    DatasetKind::Root,
+    SearchAddDataVisitor,
+    AddData,
+    Flag::ADD_DATA
+);
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -217,6 +240,63 @@ where
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+pub struct DatasetKindBasedVisitor<V, E>
+where
+    V: MetadataChainVisitor<Error = E>,
+{
+    expected_dataset_kind: DatasetKind,
+    actual_dataset_kind: DatasetKind,
+    visitor: V,
+}
+
+impl<V, E> DatasetKindBasedVisitor<V, E>
+where
+    V: MetadataChainVisitor<Error = E>,
+{
+    pub fn new(
+        expected_dataset_kind: DatasetKind,
+        actual_dataset_kind: DatasetKind,
+        visitor: V,
+    ) -> Self {
+        Self {
+            expected_dataset_kind,
+            actual_dataset_kind,
+            visitor,
+        }
+    }
+
+    pub fn into_inner(self) -> V {
+        self.visitor
+    }
+}
+
+impl<V, E> MetadataChainVisitor for DatasetKindBasedVisitor<V, E>
+where
+    V: MetadataChainVisitor<Error = E>,
+    E: Error + Send,
+{
+    type Error = E;
+
+    fn initial_decision(&self) -> Decision {
+        if self.expected_dataset_kind != self.actual_dataset_kind {
+            Decision::Stop
+        } else {
+            self.visitor.initial_decision()
+        }
+    }
+
+    #[inline]
+    fn visit(&mut self, hashed_block_ref: HashedMetadataBlockRef) -> Result<Decision, Self::Error> {
+        self.visitor.visit(hashed_block_ref)
+    }
+
+    fn finish(&self) -> Result<(), Self::Error> {
+        self.visitor.finish()
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 enum SearchSingleDataBlockVisitorKind {
     NextDataBlock,
     NextFilledNewData,
@@ -291,18 +371,30 @@ impl MetadataChainVisitor for SearchSingleDataBlockVisitor {
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 pub struct SearchActivePushSourcesVisitor {
-    hashed_blocks: Vec<(Multihash, MetadataBlockTyped<AddPushSource>)>,
+    dataset_kind: DatasetKind,
+    disabled_push_source_names: HashSet<String>,
+    active_push_sources: VecDeque<(Multihash, MetadataBlockTyped<AddPushSource>)>,
 }
 
 impl SearchActivePushSourcesVisitor {
-    pub fn new() -> Self {
+    const ROOT_SOURCE_RELATED_FLAGS: Flag = Flag::from_bits_retain(
+        Flag::SET_POLLING_SOURCE.bits()
+            | Flag::DISABLE_POLLING_SOURCE.bits()
+            | Flag::ADD_PUSH_SOURCE.bits()
+            | Flag::DISABLE_PUSH_SOURCE.bits(),
+    );
+
+    pub fn new(dataset_kind: DatasetKind) -> Self {
         Self {
-            hashed_blocks: vec![],
+            dataset_kind,
+            disabled_push_source_names: HashSet::new(),
+            active_push_sources: VecDeque::new(),
         }
     }
 
     pub fn into_hashed_blocks(self) -> Vec<(Multihash, MetadataBlockTyped<AddPushSource>)> {
-        self.hashed_blocks
+        // No reallocation
+        Vec::from(self.active_push_sources)
     }
 }
 
@@ -310,19 +402,111 @@ impl MetadataChainVisitor for SearchActivePushSourcesVisitor {
     type Error = Infallible;
 
     fn initial_decision(&self) -> Decision {
-        Decision::NextOfType(Flag::ADD_PUSH_SOURCE)
+        match self.dataset_kind {
+            DatasetKind::Root => Decision::NextOfType(Self::ROOT_SOURCE_RELATED_FLAGS),
+            DatasetKind::Derivative => Decision::Stop,
+        }
     }
 
     fn visit(&mut self, (hash, block): HashedMetadataBlockRef) -> Result<Decision, Self::Error> {
-        // TODO: Add support of `DisablePushSource` events
-        let flag = Flag::from(&block.event);
+        match &block.event {
+            MetadataEvent::SetPollingSource(_) | MetadataEvent::DisablePollingSource(_) => {
+                // > Push and polling sources are mutually exclusive.
+                // > (c) RFC-011 (https://github.com/open-data-fabric/open-data-fabric/blob/master/rfcs/011-push-ingest-sources.md#guide-level-explanation)
+                //
+                // Thus, if we encounter anything related to a polling source,
+                // it means earlier push source events are no longer relevant.
+                return Ok(Decision::Stop);
+            }
+            MetadataEvent::AddPushSource(add_push_source) => {
+                // Since we move backwards, we already know in advance whether a source is
+                // disabled.
+                let is_still_enabled = !self
+                    .disabled_push_source_names
+                    .contains(&add_push_source.source_name);
 
-        if flag == Flag::ADD_PUSH_SOURCE {
-            self.hashed_blocks
-                .push((hash.clone(), block.clone().into_typed().unwrap()));
+                if is_still_enabled {
+                    // SAFETY: block type verified
+                    let typed_block = block.clone().into_typed().unwrap();
+                    self.active_push_sources
+                        .push_front((hash.clone(), typed_block));
+                }
+            }
+            MetadataEvent::DisablePushSource(disable_push_source) => {
+                // Within this visitor, we assume the correctness of the metadata chain was
+                // already validated during event appending.
+                self.disabled_push_source_names
+                    .insert(disable_push_source.source_name.clone());
+            }
+            _ => {
+                unreachable!()
+            }
         }
 
-        Ok(Decision::NextOfType(Flag::ADD_PUSH_SOURCE))
+        Ok(Decision::NextOfType(Self::ROOT_SOURCE_RELATED_FLAGS))
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub struct SearchActivePollingSourceVisitor {
+    dataset_kind: DatasetKind,
+    hashed_block: Option<(Multihash, MetadataBlockTyped<SetPollingSource>)>,
+}
+
+impl SearchActivePollingSourceVisitor {
+    const ROOT_SOURCE_RELATED_FLAGS: Flag = Flag::from_bits_retain(
+        Flag::SET_POLLING_SOURCE.bits()
+            | Flag::DISABLE_POLLING_SOURCE.bits()
+            | Flag::ADD_PUSH_SOURCE.bits()
+            | Flag::DISABLE_PUSH_SOURCE.bits(),
+    );
+
+    pub fn new(dataset_kind: DatasetKind) -> Self {
+        Self {
+            dataset_kind,
+            hashed_block: None,
+        }
+    }
+
+    pub fn into_hashed_block(self) -> Option<(Multihash, MetadataBlockTyped<SetPollingSource>)> {
+        self.hashed_block
+    }
+}
+
+impl MetadataChainVisitor for SearchActivePollingSourceVisitor {
+    type Error = Infallible;
+
+    fn initial_decision(&self) -> Decision {
+        match self.dataset_kind {
+            DatasetKind::Root => Decision::NextOfType(Self::ROOT_SOURCE_RELATED_FLAGS),
+            DatasetKind::Derivative => Decision::Stop,
+        }
+    }
+
+    fn visit(&mut self, (hash, block): HashedMetadataBlockRef) -> Result<Decision, Self::Error> {
+        #[expect(clippy::match_same_arms)]
+        match &block.event {
+            MetadataEvent::AddPushSource(_) | MetadataEvent::DisablePushSource(_) => {
+                // > Push and polling sources are mutually exclusive.
+                // > (c) RFC-011 (https://github.com/open-data-fabric/open-data-fabric/blob/master/rfcs/011-push-ingest-sources.md#guide-level-explanation)
+                //
+                // Thus, if we encounter anything related to push sources,
+                // it means earlier pulling source events are no longer
+                // relevant.
+            }
+            MetadataEvent::SetPollingSource(_) => {
+                // SAFETY: block type verified
+                let typed_block = block.clone().into_typed().unwrap();
+                self.hashed_block = Some((hash.clone(), typed_block));
+            }
+            MetadataEvent::DisablePollingSource(_) => {
+                // No active one exists
+            }
+            _ => unreachable!(),
+        }
+
+        Ok(Decision::Stop)
     }
 }
 
@@ -418,11 +602,19 @@ where
 
 pub struct SearchKeyBlocksVisitor {
     key_blocks: Vec<HashedMetadataBlock>,
+    key_event_flags: Flag,
 }
 
 impl SearchKeyBlocksVisitor {
     pub fn new() -> Self {
-        Self { key_blocks: vec![] }
+        Self {
+            key_blocks: vec![],
+            key_event_flags: Flag::empty(),
+        }
+    }
+
+    pub fn key_event_flags(&self) -> Flag {
+        self.key_event_flags
     }
 
     pub fn into_hashed_key_blocks(self) -> Vec<HashedMetadataBlock> {
@@ -441,6 +633,7 @@ impl MetadataChainVisitor for SearchKeyBlocksVisitor {
         let flag = Flag::from(&block.event);
 
         if Flag::KEY_BLOCK.contains(flag) {
+            self.key_event_flags |= flag;
             self.key_blocks.push((hash.clone(), block.clone()));
         }