+++ CHANGELOG.md
@@ -20,6 +20,13 @@ Recommendation: for ease of reading, use the following order:
   - `AccountMut::by_ids()`.
 - Improved compatibility with EVM-based sources
 - Upgraded to `datafusion v50`
+- (#1384) `kamu inspect`: speed-up of iteration through the metadata chain.
+- (#1384) `GQL: Dataset::metadata().current_push_sources`: significant speed-up through iteration 
+  over key blocks stored in the database 
+- (#1384) General iteration optimizations based on the dataset type (root, derived). 
+  Ability to disable hints during iteration.
+- (#1384) `SearchActivePushSourcesVisitor`, `SearchActivePollingSourceVisitor`: accounting for the evolution of data 
+  sources.
 
 ## [0.248.1] - 2025-09-11
 ### Fixed